completing city, car, carType, Road, crossWay classes

+++ Project/src/model/Engine.java
@@ -4,10 +4,10 @@ package model;
  * Created by User on 12/16/2016.
  */
 public class Engine{
-    public int accelerate;
-    public int Max_Speed;
+    public double accelerate;
+    public double Max_Speed;
 
-    public Engine(int accelerate, int max_Speed){
+    public Engine(double accelerate, double max_Speed){
         this.accelerate = accelerate;
         this.Max_Speed = max_Speed;
     }
